# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/ram/all/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All With Pagination",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Get All With Pagination",
                "parameters": [
                    {
                        "description": "Get All With Pagination",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationInputWithFilter"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "RamModel response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/ram/create": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a Ram Model",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Create a Ram Model",
                "parameters": [
                    {
                        "description": "Create a RamModel",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRamModelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "RamModel response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/ram/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete RamModel from data base",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Delete RamModel",
                "responses": {
                    "204": {
                        "description": "RamModel response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/ram/get/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get Ram Model By Id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Get Ram Model By Id",
                "responses": {
                    "200": {
                        "description": "GetRamModelById response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/ram/type/{type}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get All based on ddr4 or ddr5 or ddr3",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Get All based on ddr4 or ddr5 or ddr3",
                "responses": {
                    "201": {
                        "description": "RamModel response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/v1/ram/update/{id}": {
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update RamModel",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "RamModel"
                ],
                "summary": "Update RamModel",
                "parameters": [
                    {
                        "description": " Update RamModel",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateRamModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RamModel response",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateRamModelRequest": {
            "type": "object",
            "properties": {
                "RamSize": {
                    "type": "integer"
                },
                "powerConsumption": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.Filter": {
            "type": "object",
            "properties": {
                "filterType": {
                    "type": "string"
                },
                "form": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationInputWithFilter": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/dto.Filter"
                    }
                },
                "page_number": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "sort": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SortDto"
                    }
                }
            }
        },
...
        "dto.SortDto": {
            "type": "object",
            "properties": {
                "colId": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateRamModelRequest": {
            "type": "object",
            "properties": {
                "ramSize": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "helper.Response": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "result": {},
                "statusCode": {
                    "$ref": "#/definitions/helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40002,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "Error",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError"
            ]
        }
    }
}